changed api, added event system, fully rework screen ui and adding new get point type

+++ src/main/java/ru/mentola/improvableskills/client/screen/draggable/Draggable.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public final class Draggable {
-    private double offsetY = 0, offsetX = 0;
+    private int offsetY = 0, offsetX = 0;
     private double lastMouseX = 0, lastMouseY = 0;
     private boolean isDragging;
 
@@ -18,8 +18,8 @@ public final class Draggable {
 
     public void mouseDragged(double mouseX, double mouseY, int button, double deltaX, double deltaY) {
         if (this.isDragging && button != 0) {
-            this.offsetX += (mouseX - this.lastMouseX);
-            this.offsetY += (mouseY - this.lastMouseY);
+            this.offsetX += (int) (mouseX - this.lastMouseX);
+            this.offsetY += (int) (mouseY - this.lastMouseY);
             this.lastMouseX = mouseX;
             this.lastMouseY = mouseY;
         }
